Split werewolf roles.ts into per-role files
Replaces the single roles.ts module with a roles/ directory:
- _types.ts — WerewolfRole enum, WerewolfRoleCategory, WerewolfRoleDefinition
- _config.ts — MIN_PLAYERS, MIN_ROLE_PLAYERS, defaultRoleCount
- One file per role (39 files, e.g. altruist.ts, seer.ts, werewolf-role.ts)
- index.ts — barrel assembling WEREWOLF_ROLES, re-exporting all public symbols

All existing import paths (./roles, ../roles) continue to resolve via the
directory index barrel without any changes to consumers.
