+++ src/lib/game/modes/werewolf/roles/mason.ts
@@ -1,0 +1,18 @@
+import { Team } from "@/lib/types";
+import { WakesAtNight, TargetCategory } from "../types";
+import { WerewolfRole, WerewolfRoleCategory } from "./_types";
+import type { WerewolfRoleDefinition } from "./_types";
+
+export const MASON_ROLE: WerewolfRoleDefinition = {
+  id: WerewolfRole.Mason,
+  name: "Mason",
+  summary: "Knows the identities of all other Masons",
+  description:
+    "On the first night, the Masons learn who the other Masons are. Masons have no night action after the first night, but they can trust each other completely.",
+  team: Team.Good,
+  awareOf: { roles: [WerewolfRole.Mason] },
+  wakesAtNight: WakesAtNight.FirstNightOnly,
+  targetCategory: TargetCategory.None,
+  aliases: ["brother", "freemason"],
+  category: WerewolfRoleCategory.VillagerSupport,
+};
