+++ src/lib/game/modes/werewolf/roles/pacifist.ts
@@ -1,0 +1,17 @@
+import { Team } from "@/lib/types";
+import { WakesAtNight, TargetCategory } from "../types";
+import { WerewolfRole, WerewolfRoleCategory } from "./_types";
+import type { WerewolfRoleDefinition } from "./_types";
+
+export const PACIFIST_ROLE: WerewolfRoleDefinition = {
+  id: WerewolfRole.Pacifist,
+  name: "Pacifist",
+  summary: "Always votes to acquit in daytime trials",
+  description:
+    "The Pacifist is a Good-team member who is compelled to always vote Innocent in daytime elimination trials. Their vote is automatically cast when a trial begins.",
+  team: Team.Good,
+  wakesAtNight: WakesAtNight.Never,
+  targetCategory: TargetCategory.None,
+  alwaysVotesInnocent: true,
+  category: WerewolfRoleCategory.VillagerHandicap,
+};
